сегмент номенклатури + автовибір підприємця якщо існує аналог

+++ ModelMID/DB/SegmentWares.cs
@@ -7,6 +7,6 @@ namespace ModelMID.DB
     public class SegmentWares
     {
         public Int64 CodeWares { get; set; }
-        public Int64 CodeSegment { get; set; }
+        public int CodeSegment { get; set; }
     }
 }
